requirements.txt
fastapi
uvicorn
pymongo[srv]==3.12
dnspython
cloudinary
python-dotenv
pydantic[email]
python-multipart
torch
torchaudio
torchvision
speechbrain
pyannote.audio==3.1.1
pydub
ffmpeg-python
numpy==1.24.4
huggingface-hub
requests==2.31.0
librosa